Working web application with nice UI

+++ backend/routes/favRoutes.js
@@ -3,41 +3,68 @@ const router = express.Router();
 
 const db = require('../database/db');
 
-router.post("/:routeId", (req, res) => {
-  const userId = req.user.id; // however you track users
-  const { switchId } = req.params;
+router.get("/all", (req, res) => {
+  const userId = req.headers['x-user-id'];
+  if(userId) {
+    const favorites = db.prepare("SELECT s.*, f.* FROM favorites f JOIN stations s ON f.switchId = s.stop_id WHERE f.userId = ?");
+    const rows = favorites.all(userId);
+    console.log(rows);
+    res.json({ data: rows });
+  } else {
+    res.status(404).send("NO");
+  }
+});
 
-  db.prepare(`
-    INSERT INTO favorites (userId, switchId)
-    VALUES (?, ?)
-  `).run(userId, switchId);
+router.post("/update", (req, res) => {
+  const userId = req.headers['x-user-id'];
+  if(userId) {
+    const { fid } = req.query;
 
-  res.json({ success: true });
+    db.prepare(`
+      INSERT OR IGNORE INTO favorites (userId, switchId)
+      VALUES (?, ?)
+    `).run(userId, fid);
+
+    res.json({ success: true });
+  } else {
+    res.status(404).send("NO");
+  }
 });
 
 
-router.delete("/:routeId", (req, res) => {
-  const userId = req.user.id;
-  const { switchId } = req.params;
+router.delete("/update", (req, res) => {
+  const userId = req.headers['x-user-id'];
+  if(userId) {
+    const { fid } = req.query;
 
-  db.prepare(`
-    DELETE FROM favorites
-    WHERE userId = ? AND switchId = ?
-  `).run(userId, switchId);
+    db.prepare(`
+      DELETE FROM favorites
+      WHERE userId = ? AND switchId = ?
+    `).run(userId, fid);
 
-  res.json({ success: true });
+    res.json({ success: true });
+  } else {
+    res.status(404).send("NO");
+  }
 });
 
 
-router.get("/:routeId", (req, res) => {
-  const userId = req.user.id;
-  const { switchId } = req.params;
+router.get("/get", (req, res) => {
+  const userId = req.headers['x-user-id'];
+  console.log("User id: ", userId);
+  if(userId) {
+    const { fid } = req.query;
+    
+    const favorite = db.prepare(`
+      SELECT 1 FROM favorites WHERE userId = ? AND switchId = ?
+    `).get(userId, fid);
 
-  const favorite = db.prepare(`
-    SELECT 1 FROM favorites WHERE userId = ? AND switchId = ?
-  `).get(userId, switchId);
-
-  res.json({ isFavorite: !!favorite });
+    console.log(favorite);
+    
+    res.json({ isFavorite: !!favorite });
+  } else {
+    res.status(404).send("NO");
+  }
 });
 
 module.exports = router;
